# [r2r] Create a separate top level crate and move build.rs there.