--- a/mm2src/mm2_bin_lib/src/lib.rs
+++ b/mm2src/mm2_bin_lib/src/lib.rs
@@ -1,0 +1,49 @@
+#![allow(uncommon_codepoints)]
+#![feature(async_closure)]
+#![feature(drain_filter)]
+#![feature(hash_raw_entry)]
+#![feature(integer_atomics)]
+#![recursion_limit = "512"]
+#![cfg_attr(target_arch = "wasm32", allow(unused_imports))]
+
+use mm2_core::mm_ctx::MmArc;
+use std::sync::atomic::{AtomicBool, AtomicU32, Ordering};
+#[cfg(target_arch = "wasm32")] use wasm_bindgen::prelude::*;
+
+static LP_MAIN_RUNNING: AtomicBool = AtomicBool::new(false);
+static CTX: AtomicU32 = AtomicU32::new(0);
+
+#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
+pub enum MainStatus {
+    /// MM2 is not running yet.
+    NotRunning = 0,
+    /// MM2 is running, but no context yet.
+    NoContext = 1,
+    /// MM2 is running, but no RPC yet.
+    NoRpc = 2,
+    /// MM2's RPC is up.
+    RpcIsUp = 3,
+}
+
+/// Checks if the MM2 singleton thread is currently running or not.
+pub fn mm2_status() -> MainStatus {
+    if !LP_MAIN_RUNNING.load(Ordering::Relaxed) {
+        return MainStatus::NotRunning;
+    }
+
+    let ctx = CTX.load(Ordering::Relaxed);
+    if ctx == 0 {
+        return MainStatus::NoContext;
+    }
+
+    let ctx = match MmArc::from_ffi_handle(ctx) {
+        Ok(ctx) => ctx,
+        Err(_) => return MainStatus::NoRpc,
+    };
+
+    if ctx.rpc_started.copy_or(false) {
+        MainStatus::RpcIsUp
+    } else {
+        MainStatus::NoRpc
+    }
+}