--- a/mm2src/mm2_bin_lib/src/mm2_bin.rs
+++ b/mm2src/mm2_bin_lib/src/mm2_bin.rs
@@ -1,0 +1,15 @@
+use mm2_main::mm2::mm2_main;
+
+const MM_VERSION: &str = env!("MM_VERSION");
+const MM_DATETIME: &str = env!("MM_DATETIME");
+
+#[cfg(all(target_os = "linux", target_arch = "x86_64", target_env = "gnu"))]
+#[global_allocator]
+static GLOBAL: jemallocator::Jemalloc = jemallocator::Jemalloc;
+
+fn main() {
+    #[cfg(not(target_arch = "wasm32"))]
+    {
+        mm2_main(MM_VERSION.into(), MM_DATETIME.into())
+    }
+}