--- a/mm2src/mm2_bin_lib/src/mm2_bin.rs
+++ b/mm2src/mm2_bin_lib/src/mm2_bin.rs
@@ -1,4 +1,4 @@
-use mm2_main::mm2::mm2_main;
+#[cfg(not(target_arch = "wasm32"))] use mm2_main::mm2::mm2_main;
 
 const MM_VERSION: &str = env!("MM_VERSION");
 const MM_DATETIME: &str = env!("MM_DATETIME");