--- a/mm2src/mm2_main/tests/docker_tests/swaps_confs_settings_sync_tests.rs
+++ b/mm2src/mm2_main/tests/docker_tests/swaps_confs_settings_sync_tests.rs
@@ -1,9 +1,9 @@
 use crate::generate_utxo_coin_with_random_privkey;
 use crate::integration_tests_common::enable_native;
 use common::block_on;
+use mm2_main::mm2::lp_ordermatch::OrderConfirmationsSettings;
+use mm2_main::mm2::lp_swap::get_payment_locktime;
 use mm2_test_helpers::for_tests::{mm_dump, MarketMakerIt};
-use mm2lib::mm2::lp_ordermatch::OrderConfirmationsSettings;
-use mm2lib::mm2::lp_swap::get_payment_locktime;
 use serde_json::Value as Json;
 use std::thread;
 use std::time::Duration;