--- a/mm2src/mm2_main/tests/integration_tests_common/mod.rs
+++ b/mm2src/mm2_main/tests/integration_tests_common/mod.rs
@@ -2,12 +2,12 @@ use common::executor::Timer;
 use common::log::LogLevel;
 use common::{block_on, log, now_ms};
 use crypto::privkey::key_pair_from_seed;
+use mm2_main::mm2::{lp_main, LpMainParams};
 use mm2_test_helpers::electrums::{morty_electrums, rick_electrums};
 use mm2_test_helpers::for_tests::{enable_native as enable_native_impl, init_utxo_electrum, init_utxo_status,
                                   init_z_coin_light, init_z_coin_status, MarketMakerIt};
 use mm2_test_helpers::structs::{CoinActivationResult, EnableElectrumResponse, InitTaskResult, InitUtxoStatus,
                                 InitZcoinStatus, RpcV2Response, UtxoStandardActivationResult};
-use mm2lib::mm2::{lp_main, LpMainParams};
 use serde_json::{self as json, Value as Json};
 use std::collections::HashMap;
 use std::env::var;
@@ -25,7 +25,7 @@ pub fn test_mm_start_impl() {
                 log!("test_mm_start] Starting the MarketMaker...");
                 let conf: Json = json::from_str(&conf).unwrap();
                 let params = LpMainParams::with_conf(conf).log_filter(Some(filter));
-                block_on(lp_main(params, &|_ctx| ())).unwrap()
+                block_on(lp_main(params, &|_ctx| (), "TEST".into(), "TEST".into())).unwrap()
             }
         }
     }