--- a/mm2src/mm2_main/tests/mm2_tests/orderbook_sync_tests.rs
+++ b/mm2src/mm2_main/tests/mm2_tests/orderbook_sync_tests.rs
@@ -2,6 +2,7 @@ use crate::integration_tests_common::{enable_coins_eth_electrum, enable_coins_ri
                                       enable_electrum_json, enable_z_coin_light};
 use common::{block_on, log};
 use http::StatusCode;
+use mm2_main::mm2::lp_ordermatch::MIN_ORDER_KEEP_ALIVE_INTERVAL;
 use mm2_number::{BigDecimal, BigRational, MmNumber};
 use mm2_test_helpers::electrums::rick_electrums;
 use mm2_test_helpers::for_tests::{get_passphrase, orderbook_v2, rick_conf, zombie_conf, MarketMakerIt, Mm2TestConf,
@@ -9,7 +10,6 @@ use mm2_test_helpers::for_tests::{get_passphrase, orderbook_v2, rick_conf, zombi
 use mm2_test_helpers::get_passphrase;
 use mm2_test_helpers::structs::{EnableElectrumResponse, GetPublicKeyResult, OrderbookEntryAggregate,
                                 OrderbookResponse, OrderbookV2Response, RpcV2Response, SetPriceResponse};
-use mm2lib::mm2::lp_ordermatch::MIN_ORDER_KEEP_ALIVE_INTERVAL;
 use serde_json::{self as json, json, Value as Json};
 use std::env;
 use std::str::FromStr;