--- a/mm2src/mm2_main/tests/mm2_tests_main.rs
+++ b/mm2src/mm2_main/tests/mm2_tests_main.rs
@@ -1,2 +1,4 @@
+#![cfg(not(target_arch = "wasm32"))]
+
 mod integration_tests_common;
 mod mm2_tests;