- build.rs had invalid paths to MM_VERSION and MM_DATETIME files, which caused mm2_main recompilation on every `cargo build`. Moving it to a separate crate helped to reduce recompilation times, especially for tests, which do not depend on build.rs now at all.
- Having a bit less code in `mm2_main` looks like an improvement to me as well :)
- Also fixed paths for WASM build uploads that broke after lib renaming (zip showed just warning because some other files still have the same name).
- Please note that `MANUAL_MM_VERSION=true cargo build` will turn off build.rs rerun if you have `MM_VERSION` and `MM_DATETIME` files at the project root. Though, you need to remember to set this env every time.

<details>
<summary>I have done few comparisons between old (dev) and new (this branch) compilation times</summary>
<br>
Before without file changes between builds

    cargo build --bin mm2
    Compiling mm2_main v0.1.0 (/Users/artempikulin/dev/atomicDEX-API/mm2src/mm2_main)
    Finished dev [unoptimized + debuginfo] target(s) in 21.10s

After without file changes between builds

    cargo build --bin mm2
    Compiling mm2_bin_lib v0.1.0 (/Users/artempikulin/dev/atomicDEX-API/mm2src/mm2_bin_lib)
    Finished dev [unoptimized + debuginfo] target(s) in 8.77s

Before with one file changed (lp_ordermatch.rs)

    cargo build --bin mm2
    Compiling mm2_main v0.1.0 (/Users/artempikulin/dev/atomicDEX-API/mm2src/mm2_main)
    Finished dev [unoptimized + debuginfo] target(s) in 22.92s

After with one file changed (lp_ordermatch.rs)

    cargo build --bin mm2
    Compiling mm2_bin_lib v0.1.0 (/Users/artempikulin/dev/atomicDEX-API/mm2src/mm2_bin_lib)
    Compiling mm2_main v0.1.0 (/Users/artempikulin/dev/atomicDEX-API/mm2src/mm2_main)
    Finished dev [unoptimized + debuginfo] target(s) in 15.40s

Before without file changes between builds

    cargo build --tests
    Compiling mm2_main v0.1.0 (/Users/artempikulin/dev/atomicDEX-API/mm2src/mm2_main)
    Finished dev [unoptimized + debuginfo] target(s) in 22.52s

After without file changes between builds

    cargo build --tests
    Finished dev [unoptimized + debuginfo] target(s) in 0.46s

Before with one test file (mm2src/mm2_main/tests/mm2_tests/z_coin_tests.rs) updated

    cargo build --tests
    Compiling mm2_main v0.1.0 (/Users/artempikulin/dev/atomicDEX-API/mm2src/mm2_main)
    Finished dev [unoptimized + debuginfo] target(s) in 25.22s

After with one test file (mm2src/mm2_main/tests/mm2_tests/z_coin_tests.rs) updated

    cargo build --tests
    Compiling mm2_main v0.1.0 (/Users/artempikulin/dev/atomicDEX-API/mm2src/mm2_main)
    Finished dev [unoptimized + debuginfo] target(s) in 9.68s
</details>