@ozkanonur 2022-11-24
I guess this mostly effects recompilations. At the initial compilation, most time taken is by link time optimizations. Which can be reduced using `lld`. Can you measure compilation time differences by replacing the default linker? Last time I tried I have achieved 15-25% less time on compilations by replacing default linker to `lld`.