@artemii235 2022-11-25
@ozkanonur Yeah, I have done profiling on previous week and noticed that linking takes over a minute for *release* builds. It's not a big problem actually because we compile them rarely.
On dev environments and CI, we use `dev` and `ci` profile, which don't enable lto. But recompilation still took significant time, slowing down the process. E.g. if I wanted to restart tests without changing code (to recheck unstable test or test depending on 3rd party), I would still need to wait 20+ seconds, which is quite annoying 🙂 
Please note that I will keep researching and trying to optimize overall compilation times - just preferring to do it in small iterations.